Add initial HTML structure for Abohisham-Web
I have began working with the basic in this project. I have made the “logo”, and added buttons with good looking ui, where there is hovering effect so the button becomes darker when hovering. They still do not have content though(which they will have later).

It has been a bit of time since i last coded html, therefore i have experimented a bit with css code. I have to work on the top-bar or what they now call it, because i have issue with the position of the top-bar, where it does not fill the corners of screen.