This session really drained me. It was a struggle to connect the VS-code to github, because I haven’t done it before, however, i managed to get it to work and connect through a long time of searching and pasting code in terminal.
I have also worked on fixing the code and making it more managable by making a dedicated css file for the navigation-bar. I tried to make a html file for the whole navigation-bar so i do not have to copy and paste the same code in different tabs, but i learned that i need some kind of server to run the html file. Therefore i decided to keep the code in each document.

I still got more to work on for my navigation-bar